Tweaks for NANO server.
Try to detect if we're running in a Powershell remote session, in which
case we have neither a console nor a window station.  If that's the case
it's appropriate to print the usage message to stdout rather than
attempt to display a popup.

Use EnumServicesStatusEx() instead of EnumServicesStatus().  The latter
crashes on NANO when used in the same way that works correctly on other
versions of Windows.

Thanks Kirill Kovalenko.