Support starting processes with a console on Windows 10.
By default we let the application inherit NSSM's console.  On Windows 10
Creators Update an application inheriting a console would fail to start with
error STATUS_DLL_INIT_FAILED.

Now we close our own console and call CreateProcess() with the
CREATE_NEW_CONSOLE flag if the application needs a console.

Unfortunately the ASCII art logo has to be sacrificed in the name of
progress...